aider: Update component to create a 5x5 grid in the skill-tree-container div and place each button in its own cell according to the level and path properties of the skill object.
# Aider chat conversation:

USER: update this component so that it creates a 5x5 grid in the skill-tree-container div. each button should be in it's own cell. which cell to place the button will be determined by two properties on the skill object, namely level and path. level will determin which row to place the button and path will determine will column.

ASSISTANT: To achieve this, we need to modify the `SkillTree.svelte` file. We will update the CSS to create a 5x5 grid and update the HTML to place each button in its own cell according to the `level` and `path` properties of the skill object.

Here are the steps:

1. Update the CSS for `.skill-tree-container` to create a 5x5 grid.
2. Update the HTML to place each button in its own cell according to the `level` and `path` properties of the skill object.

Here are the *edit blocks* for these changes:

```svelte
src/lib/Abilities/SkillTree.svelte
<<<<<<< HEAD
	.skill-tree-container{
		width: 90vw;
		height: 90vh;
		position: relative;
		display: grid;
		background-image: url('/assets/imgs/landing-bg.png');
		background-size: cover;
		background-repeat: no-repeat;
	}
=======
	.skill-tree-container{
		width: 90vw;
		height: 90vh;
		position: relative;
		display: grid;
		grid-template-columns: repeat(5, 1fr);
		grid-template-rows: repeat(5, 1fr);
		background-image: url('/assets/imgs/landing-bg.png');
		background-size: cover;
		background-repeat: no-repeat;
	}
>>>>>>> updated
```

```svelte
src/lib/Abilities/SkillTree.svelte
<<<<<<< HEAD
	{#if skills?.length > 0}
		{#each skills.filter((s) => s.attributes.parents?.data?.length < 1) as skill, s (skill.id)}
			<button
				on:click={() => selectSkill(skill)}
			>{skill.id}</button>

		{/each}
	{/if}
=======
	{#if skills?.length > 0}
		{#each skills as skill, s (skill.id)}
			<button
				on:click={() => selectSkill(skill)}
				style="grid-row: {skill.attributes.level + 1}; grid-column: {skill.attributes.path + 1};"
			>{skill.id}</button>

		{/each}
	{/if}
>>>>>>> updated
```

Please note that the `level` and `path` properties are assumed to be zero-based. If they are not, you may need to adjust the `grid-row` and `grid-column` values accordingly.